# Conflicts:
#	apps/api-example/schema-v1-admin.graphql
#	apps/api-example/schema-v2.graphql
#	apps/editor/src/app/routes/subscriptions/subscriptionEditView.tsx
#	libs/api/__tests__/api/private/index.ts
#	libs/api/__tests__/api/public/index.ts
#	libs/api/src/lib/graphql/subscription/subscription.private-mutation.ts
#	libs/editor/api-v2/src/lib/graphql.ts
#	libs/editor/api/src/lib/graphql.ts
#	libs/membership/api/src/lib/subscription/subscription.service.ts
#	libs/membership/editor/src/lib/subscription-flow/filter/filter-body.tsx
#	libs/peering/api/src/lib/remote/graphql.ts
#	libs/testing/src/graphql/graphql-public.ts
#	libs/testing/src/graphql/public/tag.graphql
#	libs/ui/editor/src/lib/atoms/searchAndFilter/subscriptionListFilter.tsx
#	libs/ui/editor/src/lib/blocks/teaserSlots/teaser-slots-autofill-config.tsx
#	libs/user-subscription/api/src/lib/remote-subscriptions.service.ts
#	libs/website/api/src/lib/graphql.ts

+++ libs/payment/api/src/lib/payment-provider/payrexx-subscription-payment-provider.ts
@@ -27,13 +27,6 @@ import {
   WebhookResponse,
 } from './payment-provider';
 
-export interface PayrexxSubscripionsPaymentProviderProps
-  extends PaymentProviderProps {
-  instanceName: string;
-  instanceAPISecret: string;
-  webhookSecret: string;
-}
-
 function mapPayrexxEventToPaymentStatus(event: string): PaymentState | null {
   switch (event) {
     case 'waiting':
@@ -50,16 +43,10 @@ function mapPayrexxEventToPaymentStatus(event: string): PaymentState | null {
 }
 
 export class PayrexxSubscriptionPaymentProvider extends BasePaymentProvider {
-  readonly instanceName: string;
-  readonly instanceAPISecret: string;
-  readonly webhookSecret: string;
   override readonly remoteManagedSubscription: boolean;
 
-  constructor(props: PayrexxSubscripionsPaymentProviderProps) {
+  constructor(props: PaymentProviderProps) {
     super(props);
-    this.instanceName = props.instanceName;
-    this.instanceAPISecret = props.instanceAPISecret;
-    this.webhookSecret = props.webhookSecret;
     this.remoteManagedSubscription = true;
   }
 
@@ -286,15 +273,18 @@ export class PayrexxSubscriptionPaymentProvider extends BasePaymentProvider {
       amount,
       currency,
     };
-
+    const config = await this.getConfig();
+    if (!config.apiKey || !config.payrexx_instancename) {
+      throw new Error('Payrexx missing api key or instancename');
+    }
     const signature = crypto
-      .createHmac('sha256', this.instanceAPISecret)
+      .createHmac('sha256', config.apiKey)
       .update(qs.stringify(data))
       .digest('base64');
 
     const res = await fetch(
       `https://api.payrexx.com/v1.0/Subscription/${subscriptionId}/?instance=${encodeURIComponent(
-        this.instanceName
+        config.payrexx_instancename
       )}`,
       {
         method: 'PUT',
@@ -323,12 +313,17 @@ export class PayrexxSubscriptionPaymentProvider extends BasePaymentProvider {
   }
 
   async cancelSubscriptionUpstream(subscriptionId: number) {
+    const config = await this.getConfig();
+    if (!config.apiKey || !config.payrexx_instancename) {
+      throw new Error('Payrexx missing api key or instancename');
+    }
+
     const signature = crypto
-      .createHmac('sha256', this.instanceAPISecret)
+      .createHmac('sha256', config.apiKey)
       .digest('base64');
 
     const res = await fetch(
-      `https://api.payrexx.com/v1.0/Subscription/${subscriptionId}/?instance=${this.instanceName}`,
+      `https://api.payrexx.com/v1.0/Subscription/${subscriptionId}/?instance=${config.payrexx_instancename}`,
       {
         method: 'DELETE',
         body: qs.stringify({ ApiSignature: signature }),
@@ -358,10 +353,19 @@ export class PayrexxSubscriptionPaymentProvider extends BasePaymentProvider {
     props: WebhookForPaymentIntentProps
   ): Promise<WebhookResponse> {
     const intentStates: IntentState[] = [];
+    const config = await this.getConfig();
 
     // Protect endpoint
     const apiKey = props.req.query['apiKey'] as string;
-    if (!this.timeConstantCompare(apiKey, this.webhookSecret)) {
+    if (
+      !this.timeConstantCompare(
+        apiKey,
+        this.assertProperty(
+          'webhookEndpointSecret',
+          config.webhookEndpointSecret
+        )
+      )
+    ) {
       return {
         status: 403,
         message: 'Invalid Api Key',
